requirements.txt
discord.py==1.5.1
dnspython==1.16.0
PyNaCl==1.3.0
async-timeout==3.0.1
httpx==0.16.*
bs4==0.0.1
asyncpg
markdown
jishaku
requests
python-chess
svglib